Clean up counter().
It is a bad practice to catch all exceptions because you will
probably catch things like typos which should really be raised
to the top level.

This commit changes the behavior of the function to return the actual
value of the card in blackjack.  I'm not sure why it was returning 0,
-1, and 1 before, but I don't know that much about blackjack.